Update README.md
Here is the start of a thread on a proposal for a Website. I think it’d be nice to have a few good responsive site before a release, so that people’s first experience/exposure to the project/docs is very positive, and they leave with ideas about how it might eventually fit into other stuff they’re building.